+++ src/com/magento/idea/magento2plugin/actions/generation/generator/util/GetCodeTemplate.java
@@ -21,6 +21,7 @@ public class GetCodeTemplate {
 
         return INSTANCE;
     }
+
     GetCodeTemplate (Project project) {
         this.project = project;
     }