+++ src/com/magento/idea/magento2plugin/magento/files/ModuleDiXml.java
@@ -10,11 +10,19 @@ import com.intellij.lang.xml.XMLLanguage;
 public class ModuleDiXml implements ModuleFileInterface {
     public static String FILE_NAME = "di.xml";
     public static String TEMPLATE = "Magento Module DI Xml";
+
+    //code templates
     public static String TEMPLATE_PLUGIN = "Magento Module DI Xml Plugin";
+    public static String TEMPLATE_PREFERENCE = "Magento Module DI Xml Preference";
+
+    //tags
     public static String PLUGIN_TYPE_TAG = "type";
     public static String PLUGIN_TYPE_ATTRIBUTE = "type";
     public static String PLUGIN_TAG_NAME = "plugin";
     public static String PLUGIN_TYPE_ATTR_NAME = "name";
+    public static String PREFERENCE_TAG_NAME = "preference";
+    public static String PREFERENCE_ATTR_FOR = "for";
+
     private static ModuleDiXml INSTANCE = null;
 
     public static ModuleDiXml getInstance() {